Verify there are no students enrolled on the orientation / section.

+++ app/Http/Controllers/OrientationController.php
@@ -178,6 +178,11 @@ class OrientationController extends Controller
             return redirect(route('dashboard'))->with('errMessage', 'This orientation could not be deleted because it contains sections. Please delete all sections and try again.');
         }
 
+        // Check for enrolled students.
+        if ($orientation->users->count() > 0) {
+            return redirect(route('dashboard'))->with('errMessage', 'This orientation could not be deleted because there are students currently enrolled. Please unenroll any students before you delete this orientation.');
+        }
+
         DB::table('orientations')->delete($orientation->id);
 
         return redirect(route('dashboard'))->with('message', 'Orientation deleted.');
